feat(test): separate tests
