+++ packages/generator/test/__utils__/index.js
@@ -1,0 +1,14 @@
+export const createNuxt = () => ({
+  ready: jest.fn(),
+  callHook: jest.fn(),
+  server: {
+    renderRoute: jest.fn(() => ({ html: 'rendered html' }))
+  },
+  options: {
+    srcDir: '/var/nuxt/src',
+    buildDir: '/var/nuxt/build',
+    generate: { dir: '/var/nuxt/generate' },
+    build: { publicPath: '__public' },
+    dir: { static: '/var/nuxt/static' }
+  }
+})
